pacman-static: nvchecker port over to v2
Which for inexplicable reasons decided to inflict the ugliness of *.toml
on me. So yay for pointlessly reformatting every line in the file, and
making it twice as verbose and half as readable.

This would be livable if there were any rationale offered for why the
switch was made, but @lilydjwg has merely noted in the documentation "it
has been changed because I said so, here's a script to migrate your
config file and totally fail to produce something usable".

This limited config is so simple it doesn't actually break, fortunately.